+++ project-addons/pmp_landed_costs/models/stock_landed_costs.py
@@ -4,6 +4,7 @@
 from odoo import models, api, _, fields, tools
 import odoo.addons.decimal_precision as dp
 from odoo.exceptions import UserError
+import datetime
 
 
 class StockLandedCost(models.Model):
@@ -231,3 +232,98 @@ class StockLandedCostLines(models.Model):
                 self.product_id.property_stock_account_input.id or \
                 self.product_id.categ_id.\
                 property_stock_account_input_categ_id.id
+
+
+class LandedCostCreator(models.TransientModel):
+    """
+    Models the creation of stock_landed_costs from import_sheet.
+    Shows a list with all products with no weight
+    """
+    _name = 'landed.cost.creator.wizard'
+
+    import_sheet_id = fields.Many2one('import.sheet', string='Import sheet')
+    product_ids = fields.Many2many('product.product', string='Products')
+    container_id = fields.Many2one(related='import_sheet_id.container_id')
+
+    def _get_account_journal_for_landed_cost(self):
+        """
+        Returns the correct account journal to assign to landed cost
+
+        Return:
+        ------
+        account.journal
+        """
+        # FIXME:
+        # self.env['account.journal'].search([()])
+        return 1
+
+    def _get_product_for_landed_cost_line(self):
+        """
+        Returns the correct product to assign to landed cost lines
+
+        Return:
+        ------
+        product.product
+        """
+        # FIXME:
+        # self.env['product.product'].search([()])
+        return 2724
+
+    def _get_account_for_landed_cost_line(self):
+        """
+        Returns the correct account to assign to landed cost lines
+
+        Return:
+        ------
+        account.account
+        """
+        # FIXME:
+        # self.env['account.journal'].search([()])
+        return 845
+
+    def create_landed_cost(self):
+        """
+        Creates landed cost associated to import_sheet_id.
+        This landed cost has two cost lines.
+        """
+        landed_cost = self.env['stock.landed.cost'].create({
+            'date': datetime.date.today(),
+            'picking_ids': [(6, 0, self.container_id.picking_ids.ids)],
+            'container_ids': [(4, self.container_id.id)],
+            'account_journal_id': self._get_account_journal_for_landed_cost(),
+            'forwarder_invoice': self.import_sheet_id.forwarder_comercial,
+            'import_sheet_id': self.import_sheet_id.id
+        })
+        self._create_cost_lines(landed_cost)
+        return
+
+    def _create_cost_lines(self, landed_cost):
+        """
+        Creates two stock.landed.cost.lines.
+        The first by fee, the second by destination costs
+
+        Parameters:
+        ----------
+        landed_cost: stock.landed.cost
+            Landed cost where we are going to create lines
+        """
+        create_line = self.env['stock.landed.cost.lines'].create
+        product_id = self._get_product_for_landed_cost_line()
+        account_id = self._get_account_for_landed_cost_line()
+        create_line({
+            'cost_id': landed_cost.id,
+            'product_id': product_id,
+            'name': 'Arancel',
+            'account_id': account_id,
+            'split_method': 'by_tariff',
+            'price_unit': self.import_sheet_id.calculate_fee_price()
+        })
+        create_line({
+            'cost_id': landed_cost.id,
+            'product_id': product_id,
+            'name': 'Coste en destino',
+            'account_id': account_id,
+            'split_method': 'equal',
+            'price_unit': self.import_sheet_id.calculate_destination_cost_price()
+        })
+        return
